refactor(sync): serve connectors, projects, plugins, skills, and accounts from the runtime
These routes ran in-process in the Next server, importing
@local-studio/agent-runtime modules directly, so pointing the frontend at a
remote runtime (LOCAL_STUDIO_AGENT_RUNTIME_URL) still read and wrote
connector, grant, project, plugin, skill, prompt-template, and Google
account state on the machine serving the UI - split-brain with the runtime
actually running the agent. The runtime now owns these endpoints
(connector-handlers, google-account-handlers, project-handlers,
plugin-handlers, discovery-handlers, all registered in http/app.ts, with
prompt-templates-store moved into the runtime package), and the frontend
routes are thin proxyToAgentRuntime passthroughs that keep their
requireApiAccess guards. URL shapes and response bodies are unchanged -
handler logic moved verbatim, so the UI cannot tell the difference.
