+++ frontend/src/app/api/agent/accounts/google/authorize/route.ts
@@ -1,58 +1,18 @@
-import { NextResponse, type NextRequest } from "next/server";
-import { Effect, Schema } from "effect";
-import { GoogleAccountError } from "@local-studio/agent-runtime/google-account";
-import {
-  beginGoogleLoopbackAuthorization,
-  cancelGoogleLoopbackAuthorization,
-} from "@local-studio/agent-runtime/google-oauth-loopback";
+import { type NextRequest } from "next/server";
 import { requireApiAccess } from "@/lib/auth/guard";
+import { proxyToAgentRuntime } from "@/app/api/agent/proxy-to-runtime";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
 
-const GoogleAccountInputSchema = Schema.Struct({
-  account: Schema.Union([Schema.Literal("gmail"), Schema.Literal("google-calendar")]),
-});
-
-export async function POST(request: NextRequest) {
+export async function POST(request: NextRequest): Promise<Response> {
   const denied = requireApiAccess(request);
   if (denied) return denied;
-  let input: typeof GoogleAccountInputSchema.Type;
-  try {
-    input = Schema.decodeUnknownSync(GoogleAccountInputSchema)(await request.json());
-  } catch {
-    return NextResponse.json({ error: "account is required" }, { status: 400 });
-  }
-  try {
-    return NextResponse.json(
-      await Effect.runPromise(beginGoogleLoopbackAuthorization(input.account)),
-    );
-  } catch (error) {
-    const status = error instanceof GoogleAccountError ? error.status : 500;
-    return NextResponse.json(
-      { error: error instanceof Error ? error.message : "Google sign-in failed" },
-      { status },
-    );
-  }
+  return proxyToAgentRuntime(request);
 }
 
-export async function DELETE(request: NextRequest) {
+export async function DELETE(request: NextRequest): Promise<Response> {
   const denied = requireApiAccess(request);
   if (denied) return denied;
-  let input: typeof GoogleAccountInputSchema.Type;
-  try {
-    input = Schema.decodeUnknownSync(GoogleAccountInputSchema)(await request.json());
-  } catch {
-    return NextResponse.json({ error: "account is required" }, { status: 400 });
-  }
-  try {
-    await Effect.runPromise(cancelGoogleLoopbackAuthorization(input.account));
-    return NextResponse.json({ cancelled: true });
-  } catch (error) {
-    const status = error instanceof GoogleAccountError ? error.status : 500;
-    return NextResponse.json(
-      { error: error instanceof Error ? error.message : "Google sign-in cancellation failed" },
-      { status },
-    );
-  }
+  return proxyToAgentRuntime(request);
 }
