+++ frontend/src/app/api/agent/connectors/ssh-server-path/route.ts
@@ -1,12 +1,12 @@
-import { NextResponse, type NextRequest } from "next/server";
-import { resolveBundledMcpServerPath } from "@local-studio/agent-runtime/pi-runtime-helpers";
+import { type NextRequest } from "next/server";
 import { requireApiAccess } from "@/lib/auth/guard";
+import { proxyToAgentRuntime } from "@/app/api/agent/proxy-to-runtime";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
 
-export async function GET(request: NextRequest) {
+export async function GET(request: NextRequest): Promise<Response> {
   const denied = requireApiAccess(request);
   if (denied) return denied;
-  return NextResponse.json({ path: resolveBundledMcpServerPath("ssh-remote.mjs") });
+  return proxyToAgentRuntime(request);
 }
