+++ frontend/src/app/api/agent/prompt-templates/load/route.ts
@@ -1,12 +1,9 @@
-import { NextRequest, NextResponse } from "next/server";
-import { loadPromptTemplateInstructions } from "@/features/agent/prompt-templates-store";
+import { type NextRequest } from "next/server";
+import { proxyToAgentRuntime } from "@/app/api/agent/proxy-to-runtime";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
 
-export async function GET(request: NextRequest) {
-  const templatePath = request.nextUrl.searchParams.get("path") ?? "";
-  const template = templatePath ? loadPromptTemplateInstructions(templatePath) : null;
-  if (!template) return NextResponse.json({ error: "Template not found" }, { status: 404 });
-  return NextResponse.json({ template });
+export async function GET(request: NextRequest): Promise<Response> {
+  return proxyToAgentRuntime(request);
 }
