+++ frontend/src/app/api/agent/prompt-templates/route.ts
@@ -1,9 +1,9 @@
-import { NextResponse } from "next/server";
-import { discoverPromptTemplates } from "@/features/agent/prompt-templates-store";
+import { type NextRequest } from "next/server";
+import { proxyToAgentRuntime } from "@/app/api/agent/proxy-to-runtime";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
 
-export async function GET() {
-  return NextResponse.json({ templates: discoverPromptTemplates() });
+export async function GET(request: NextRequest): Promise<Response> {
+  return proxyToAgentRuntime(request);
 }
