+++ frontend/src/app/api/agent/skills/load/route.ts
@@ -1,12 +1,9 @@
-import { NextRequest, NextResponse } from "next/server";
-import { loadSkillInstructions } from "@local-studio/agent-runtime/skill-discovery";
+import { type NextRequest } from "next/server";
+import { proxyToAgentRuntime } from "@/app/api/agent/proxy-to-runtime";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
 
-export async function GET(request: NextRequest) {
-  const skillPath = request.nextUrl.searchParams.get("path") ?? "";
-  const skill = skillPath ? loadSkillInstructions(skillPath) : null;
-  if (!skill) return NextResponse.json({ error: "Skill not found" }, { status: 404 });
-  return NextResponse.json({ skill });
+export async function GET(request: NextRequest): Promise<Response> {
+  return proxyToAgentRuntime(request);
 }
